requirements.txt
Flask==2.3.3
requests==2.31.0
google-generativeai==0.8.5
google-cloud-speech==2.33.0
python-dotenv==1.0.0
yt-dlp==2025.7.21
pydub==0.25.1
gunicorn==21.2.0 